requirements.txt
APScheduler==3.7.0
avro==1.10.1
backports.entry-points-selectable==1.1.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colored==1.4.2
crcmod==1.7
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
decorator==4.4.2
dill==0.2.6
distlib==0.3.2
filelock==3.0.12
Flask==1.1.2
Flask-Script==2.0.6
future==0.16.0
gapic-google-cloud-pubsub-v1==0.15.4
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-bigquery==0.25.0
google-cloud-core==0.25.0
google-cloud-pubsub==0.26.0
google-gax==0.15.16
googleapis-common-protos==1.5.5
googledatastore==7.0.1
grpc-google-iam-v1==0.11.4
grpcio==1.34.0
httplib2==0.9.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==0.17.0
kaggle==1.5.10
kiwisolver==1.3.1
lmdb==1.1.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mock==2.0.0
networkx==2.5
numpy==1.19.3
oauth2client==3.0.0
opencv-python==4.5.1.48
opfunu==0.8.0
optproblems==1.3
pandas==1.1.4
pbr==5.5.1
Pillow==8.0.1
platformdirs==2.0.2
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.3.0
pthflops==0.4.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyreadline==2.1
pyshp==2.1.2
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
PyYAML==3.13
requests==2.25.0
rsa==4.6
scikit-learn==0.24.1
scipy==1.5.4
Shapely==1.7.1
six==1.10.0
speedtest-cli==2.1.2
spotipy==2.16.1
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.59.0
typing==3.6.6
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.5.0
Werkzeug==1.0.1
